requirement.txt
absl-py==2.1.0
archspec @ file:///croot/archspec_1709217642129/work
astunparse==1.6.3
boltons @ file:///croot/boltons_1677628692245/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cloudpickle==3.0.0
conda @ file:///croot/conda_1722004606466/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
distro @ file:///croot/distro_1714488253808/work
einops==0.8.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2024.6.1
fvcore==0.1.5.post20221221
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.5
h5py==3.11.0
huggingface-hub==0.24.6
idna @ file:///croot/idna_1714398848350/work
imageio==2.35.1
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MedPy==0.5.2
menuinst @ file:///croot/menuinst_1723567589013/work
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
natsort==8.4.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.2
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
portalocker==2.10.1
progressbar2==4.4.2
protobuf==4.25.4
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.2
requests @ file:///croot/requests_1721410876868/work
rich==13.7.1
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
SimpleITK==2.4.0
sip==6.8.6
six==1.16.0
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorlayer==2.2.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==1.0.8
tomli==2.0.1
torch==2.4.0
torchinfo==1.8.0
torchvision==0.19.0
tqdm @ file:///croot/tqdm_1716395931952/work
triton==3.0.0
truststore @ file:///croot/truststore_1695244293384/work
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1718912636303/work
Werkzeug==3.0.3
wrapt==1.16.0
xlrd==2.0.1
yacs==0.1.8
zstandard @ file:///croot/zstandard_1714677652653/work
